% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plots.R
\name{plotFCs}
\alias{plotFCs}
\alias{plotRs}
\title{Scatterplot of fold changes.}
\usage{
plotFCs(object, ...)

plotRs(object, ...)

\S4method{plotFCs}{xtailResults}(
  object,
  log2FC.cutoff = 1,
  cex = 1,
  xlim,
  ylim,
  ...,
  cex.lab,
  cex.axis,
  cex.main,
  cex.sub,
  pch
)
}
\arguments{
\item{object}{a xtailResults}
}
\description{
A simple function that plots the fold change in mRNA and RPF.
Different classes of genes are labeled with different of color.
Each dot, representing a particular gene, are color coded by P-value (-log10)
of its differential translation across two conditions, which was estimated with xtail.
}
