% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRs}
\alias{plotRs}
\title{Scatterplot of log2 RPF-to-mRNA ratios.}
\usage{
\S4method{plotRs}{xtailResults}(
  object,
  log2R.cutoff = 1,
  cex = 1,
  xlim,
  ylim,
  ...,
  cex.lab,
  cex.axis,
  cex.main,
  cex.sub,
  pch
)
}
\arguments{
\item{object}{a xtailResults}
}
\description{
A simple function that plots the log2 RPF-to-mRNA ratios in two conditions.
Different classes of genes are labeled with different of color.
Each dot, representing a particular gene, are color coded by P-value (-log10)
}
