% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{resultsTable}
\alias{resultsTable}
\alias{resultsTable.xtailResults}
\title{Results table of xtail results
Print the results of xtail}
\usage{
resultsTable(object, ...)

\method{resultsTable}{xtailResults}(object, sort.by)
}
\arguments{
\item{object}{a \code{xtailResults} obect}

\item{sort.by}{the sorted column. By default, the pvalue.adjust is used.}
}
\description{
Results table of xtail results
Print the results of xtail
}
