% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary.xtailResults}
\title{Summarize xtail results
Print a summary of the results of xtail}
\usage{
\method{summary}{xtailResults}(object,alpha)
}
\arguments{
\item{object}{a \code{xtailResults} obect}

\item{alpha}{the adjusted p-value cutoff. If not set,
the default is 0.1.}
}
\description{
Summarize xtail results
Print a summary of the results of xtail
}
