% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{volcano plot}
\usage{
\S4method{volcanoPlot}{xtailResults}(object)
}
\arguments{
\item{object}{a xtailResults}
}
\description{
A simple function that plots log2 fold change of TE and -log10 of pvalues
The genes are color coded by P-value (-log10)
}
