% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtail-package.R
\docType{package}
\name{xtail-package}
\alias{xtail-package}
\title{A tool to quantitatively assess differential translations with ribosome profiling data.}
\description{
By pairwise comparisons of ribosome profiling data, Xtail
identifies differentially translated genes across two experimental or
physiological conditions.
}
\references{
Zhengtao Xiao, Qin Zou, Yu Liu, and Xuerui Yang: Genome-wide
assessment of differential translations with ribosome profiling data.
}
\author{
Zhengtao xiao
}
