% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtail-package.R
\docType{data}
\name{xtaildata}
\alias{xtaildata}
\title{xtail example data}
\format{
a list of matrices
}
\usage{
data(xtestdata)
}
\description{
The \code{xtail} package includes some example data
}
\keyword{datasets}
