% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fc.R
\name{plotFCs}
\alias{plotFCs}
\alias{plotFCs,xtail-method}
\title{Scatterplot of fold changes.}
\usage{
plotFCs(object, ...)

\S4method{plotFCs}{xtail}(object, log2FC.cutoff = 1, xlim = NULL, ylim = NULL)
}
\arguments{
\item{object}{a \code{xtail} object}

\item{...}{optional arguments. Currently not used}

\item{log2FC.cutoff}{cutoff for categorizing results}

\item{xlim}{argument for limiting the range plotted on the x axis. See
\code{\link[ggplot2:scale_continuous]{scale_continuous}} for more
details.}

\item{ylim}{argument for limiting the range plotted on the y axis. See
\code{\link[ggplot2:scale_continuous]{scale_continuous}} for more
details.}
}
\value{
a \code{ggplot} object
}
\description{
A simple function that plots the fold change in mRNA and RPF.
Different classes of genes are labeled with different of color.
Each dot, representing a particular gene, are color coded by the category of
expression change.
}
\examples{
data(xtailres)
plotFCs(xtailres)
}
