\name{plotFCs}
\alias{plotFCs}
\docType{package}
\title{
Scatter plot of the log2 fold changes of mRNA and RPF
}
\description{
A simple function that plots the log2 fold changes of mRNA and RPF
}
\usage{
plotFCs(object, log2FC.cutoff = 1, cex=1, xlim=NA, ylim=NA, ...)
}
\arguments{
  \item{object}{a object that generated by \code{xtail} function}
  \item{log2FC.cutoff}{the threshold of log2 fold change of mRNA and RPF}
  \item{cex}{cex}
  \item{xlim}{the bound for points on the x-axis}
  \item{ylim}{the bound for points on the y-axis}
  \item{...}{further arguments to 'plot'}
}
\examples{
  #load data
  data(xtaildata)
  #Get the mrna count data and rpf count data
  test.mrna <- xtaildata$mrna
  test.rpf <- xtaildata$rpf
  #Assign condition labels to samples.
  condition <- c("control","control","treat","treat")
  #run xtail
  test.results <- xtail(test.mrna,test.rpf,condition)
  #log2 fold changes plot
  plotFCs(test.results)
}
\author{
Zhengtao xiao
}
\references{
Zhengtao Xiao, Qin Zou, Yu Liu, and Xuerui Yang: Genome-wide assessment of differential translations with ribosome profiling data.
}
