% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-r.R
\name{plotRs}
\alias{plotRs}
\alias{plotRs,xtail-method}
\title{Scatterplot of log2 RPF-to-mRNA ratios.}
\usage{
plotRs(object, ...)

\S4method{plotRs}{xtail}(object, log2R.cutoff = 1, xlim = NULL, ylim = NULL)
}
\arguments{
\item{object}{a \code{xtail} object}

\item{...}{optional arguments. Currently not used}

\item{log2R.cutoff}{cutoff for categorizing results}

\item{xlim}{argument for limiting the range plotted on the x axis. See
\code{\link[ggplot2:scale_continuous]{scale_continuous}} for more
details.}

\item{ylim}{argument for limiting the range plotted on the y axis. See
\code{\link[ggplot2:scale_continuous]{scale_continuous}} for more
details.}
}
\description{
A simple function that plots the log2 RPF-to-mRNA ratios in two conditions.
Different classes of genes are labeled with different of color.
Each dot, representing a particular gene, are color coded by the category of
expression change.
}
