\name{resultsTable}
\alias{resultsTable}
\docType{package}
\title{
Extract the results table from the object generated by xtail
}
\description{
This function extracts the results of the xtail
}
\usage{
resultsTable(object,sort.by="pvalue.adjust", log2FCs = FALSE, log2Rs = FALSE)
}
\arguments{
  \item{object}{a object that generated by \code{xtail} function}
  \item{sort.by}{the column name that used for sorting the result table, such as "log2FC_TE_v1", "log2FC_TE_v2" ,"pvalue_v1" , "pvalue_v2", "pvalue_final","pvalue.adjust"}
  \item{log2FCs}{report the "mRNA_log2FC" and "PRF_log2FC" of first pipline of \code{xtail}}
  \item{log2Rs}{report the log2 Ratios of second pipline of \code{xtail}}
}
\examples{
  #load data
  data(xtaildata)
  #Get the mrna count data and rpf count data
  test.mrna <- xtaildata$mrna
  test.rpf <- xtaildata$rpf
  #Assign condition labels to samples.
  condition <- c("control","control","treat","treat")
  #run xtail
  test.results <- xtail(test.mrna,test.rpf,condition)
  #extract the results table
  test.tab <- resultsTable(test.results)
  #write the results
  write.table(test.tab,"test_results.txt",quote=F,sep="\t")
}
\author{
Zhengtao xiao
}
\references{
Zhengtao Xiao, Qin Zou, Yu Liu, and Xuerui Yang: Genome-wide assessment of differential translations with ribosome profiling data.
}
