% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runXtail.R
\name{runXtail}
\alias{runXtail}
\alias{runXtail,SummarizedExperiment-method}
\alias{addXtail}
\alias{addXtail,SummarizedExperiment-method}
\title{xtail analysis wrapper for \code{SummarizedExperiment} objects}
\usage{
runXtail(x, mrna_assay, rpf_assay, ...)

\S4method{runXtail}{SummarizedExperiment}(x, mrna_assay, rpf_assay, ...)

addXtail(x, ...)

\S4method{addXtail}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{a \code{RpfSummarizedExperiment} object}

\item{mrna_assay}{a scalar character. The name of the \code{assay} containing
the \code{mRNA} data.}

\item{rpf_assay}{a scalar character. The name of the \code{assay} containing
the \code{rpf} data.}

\item{...}{\itemize{
\item{For \code{runXtail}: additional parameters passed on to
\code{\link[=xtail]{xtail}}.}
\item{For \code{addXtail}: additional parameters passed on to
\code{runXtail} and \code{\link[=xtail]{xtail}}.}
}}
}
\value{
A \code{\link[=xtail]{xtail}} object for \code{runXtail} or
an object of \code{class(x)}.
}
\description{
The \code{\link[=xtail]{xtail}} function can be used directly with
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
objects. The \code{mrna} and \code{rpf} data must be stored as two separate
assays.
}
\details{
See \code{\link[=xtail]{xtail}} for more details on the analysis
function.
}
\examples{
data(xtaildata)
test.mrna <- xtaildata$mrna[1:100,]
test.rpf <- xtaildata$rpf[1:100,]
condition <- c("control","control","treat","treat")

se <- SummarizedExoeriment(assays = list(mrna = test.mrna, rpf = test.rpf),
                           colData = DataFrame(condition = condition))
xtail <- runXtail(se, "mrna", "rpf", bins = 1000, threads = 2)
xtail

se <- addXtail(se, "mrna", "rpf", bins = 1000, threads = 2)
rowData(se)
}
