\name{summary}
\alias{summary}
\docType{package}
\title{
Summarize Xtail results
}
\description{
Print a summary of the results from a xtail analysis.
}
\usage{
summary(object,alpha)
}
\arguments{
  \item{object}{a object that generated by \code{xtail} function}
  \item{alpha}{the adjusted p-value cutoff, default is 0.1}
}
\examples{
  #load data
  data(xtaildata)
  #Get the mrna count data and rpf count data
  test.mrna <- xtaildata$mrna
  test.rpf <- xtaildata$rpf
  #Assign condition labels to samples.
  condition <- c("control","control","treat","treat")
  #run xtail
  test.results <- xtail(test.mrna,test.rpf,condition)
  #print the summary of the xtail results
  summary(test.results)
}
\author{
Zhengtao xiao
}
\references{
Zhengtao Xiao, Qin Zou, Yu Liu, and Xuerui Yang: Genome-wide assessment of differential translations with ribosome profiling data.
}
