% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{volcanoPlot}
\alias{volcanoPlot}
\alias{volcanoPlot,xtail-method}
\title{Volcano plot}
\usage{
volcanoPlot(object, ...)

\S4method{volcanoPlot}{xtail}(object)
}
\arguments{
\item{object}{a \code{xtail} object}

\item{...}{optional arguments. Currently not used}
}
\description{
A simple function that plots log2 fold change of TE and -log10 of pvalues
The genes are color coded by P-value (-log10)
}
