\name{volcanoPlot}
\alias{volcanoPlot}
\docType{package}
\title{
Scatter plot of the log2 fold changes of TE and p-values
}
\description{
A simple function that plots the log2 fold changes of TE and negative log10 p-valules
}
\usage{
volcanoPlot(object)
}
\arguments{
  \item{object}{a object that generated by \code{xtail} function}
}
\examples{
  #load data
  data(xtaildata)
  #Get the mrna count data and rpf count data
  test.mrna <- xtaildata$mrna
  test.rpf <- xtaildata$rpf
  #Assign condition labels to samples.
  condition <- c("control","control","treat","treat")
  #run xtail
  test.results <- xtail(test.mrna,test.rpf,condition)
  #log2 fold changes plot
  volcanoPlot(test.results)
}
\author{
Zhengtao xiao
}
\references{
Zhengtao Xiao, Qin Zou, Yu Liu, and Xuerui Yang: Genome-wide assessment of differential translations with ribosome profiling data.
}
