% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtail-class.R
\name{write.xtail}
\alias{write.xtail}
\title{Write xtail results as table}
\usage{
write.xtail(object, ..., write.args = list())
}
\arguments{
\item{object}{a \code{xtail} object}

\item{...}{arguments passed onto \code{\link[=resultsTable]{resultsTable}}}

\item{write.args}{a list of arguments passed onto \code{write.table}}
}
\value{
invisible result from \code{write.table}
}
\description{
xtail results can be directly written to file using \code{write.xtail}
}
\examples{
data(xtailres)
write.xtail(xtailres, file = tempfile())
}
\seealso{
\code{\link[utils:write.table]{write.table}}
}
