% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{write.xtail}
\alias{write.xtail}
\alias{write.table}
\title{write xtail results as table}
\usage{
\method{write.xtail}{xtailResults}(object)
}
\arguments{
\item{object}{a \code{xtailResults} obect}
}
\description{
write xtail results as table
}
