\name{write.xtail}
\alias{write.xtail}
\docType{package}
\title{
Write the xtail results table to file
}
\description{
A simple function that write the xtail results to a file
}
\usage{
write.xtail(object,sort.by="pvalue.adjust", log2FCs = FALSE, log2Rs = FALSE, ...)
}
\arguments{
  \item{object}{a object that generated by \code{xtail} function}
  \item{sort.by}{the column name that used for sorting the result table, such as "log2FC_TE_v1", "log2FC_TE_v2" ,"pvalue_v1" , "pvalue_v2", "pvalue_final","pvalue.adjust"}
  \item{log2FCs}{report the "mRNA_log2FC" and "PRF_log2FC" of first pipline of \code{xtail}}
  \item{log2Rs}{report the log2 Ratios of second pipline of \code{xtail}}
  \item{...}{further arguments to 'write.table'}
}
\examples{
  #load data
  data(xtaildata)
  #Get the mrna count data and rpf count data
  test.mrna <- xtaildata$mrna
  test.rpf <- xtaildata$rpf
  #Assign condition labels to samples.
  condition <- c("control","control","treat","treat")
  #run xtail
  test.results <- xtail(test.mrna,test.rpf,condition)
  #write the results to file
  write.xtail(test.results,"test_results.txt",quote=F,sep="\t")
}
\author{
Zhengtao xiao
}
\references{
Zhengtao Xiao, Qin Zou, Yu Liu, and Xuerui Yang: Genome-wide assessment of differential translations with ribosome profiling data.
}
