% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtail-class.R
\name{xtail-accessors}
\alias{xtail-accessors}
\alias{resultsTable}
\alias{conditions,xtail-method}
\alias{resultsNum}
\alias{resultsNum,xtail-method}
\alias{resultsTable,xtail-method}
\alias{summary.xtail}
\title{Results table of xtail results}
\usage{
\S4method{conditions}{xtail}(object)

resultsNum(object, ...)

\S4method{resultsNum}{xtail}(object, ...)

resultsTable(object, ...)

\S4method{resultsTable}{xtail}(object, sort.by = NULL, log2FCs = FALSE, log2Rs = FALSE)

summary.xtail(object, alpha = 0.1, ...)
}
\arguments{
\item{object}{a \code{xtail} object}

\item{...}{optional arguments. Currently not used}

\item{sort.by}{the column to sort with. Defaults to \code{NULL} to disable
sorting.}

\item{log2FCs}{\code{TRUE} or \code{FALSE}: Should log2 fold change values be
returned? (defaults to \code{TRUE})}

\item{log2Rs}{\code{TRUE} or \code{FALSE}: Should log2 ratio values be
returned? (defaults to \code{TRUE})}

\item{alpha}{cut off for summarizing results. Only results with a adjusted
p-value lower than \code{alpha} will be reported.}
}
\value{
a \code{DataFrame} with the results or numeric vectors
}
\description{
To retrieve the results from the xtail run use one of the accessor functions.
}
\examples{
data(xtailres)
resultsTable(xtailres)
conditions(xtailres)
resultsNum(xtailres)

# sorting or results
resultsTable(xtailres, sort.by = "pvalue.adjust")
}
