\name{testdata}
\alias{testdata}
\docType{data}
\title{Sample mRNA and RPF data}
\description{
  Sample mRNA and RPF data for 11391 genes and 4 samples.
}
\usage{data(testdata)}
\format{
  A list containing two data frames, mrna that has counts data of mRNA samples and rpf that has counts data of RPF samples.
}
\keyword{datasets}

